requirements.txt
pywebview==4.4.1
pystray==0.19.5
pillow==10.2.0
keyboard==0.13.5
PyAudio==0.2.14
openai==1.12.0
pyperclip==1.8.2
pyinstaller==6.3.0 